Added confirmation popup v0.3

+++ src/main/webapp/WEB-INF/jsp/listBacktraces.jsp
@@ -159,7 +159,7 @@
 				        	<p>
 				        		<a href="viewBacktrace/${backtrace.id}.html">view</a> | 
 				        		<a href="editBacktraceForm/${backtrace.id}.html">edit</a> |
-								<a href="deleteBacktrace/${backtrace.id}.html">delete</a>
+								<a href="deleteBacktrace/${backtrace.id}.html" onClick="return confirmAction()">delete</a>
 							</p>				        
 				        </td>
 				        <td>${backtrace.name}</td>
@@ -189,6 +189,10 @@
 		<script>
 			$('#myTable').tooltip({'trigger':'hover', 'title': 'To filter/search for more than one word use logic operators: e.g. in case of three words per entry use "one && two && three" or "one AND two AND three"!'});
 			$('#reset').tooltip({'trigger':'hover', 'title': 'Press RESET to clear all filtering/searching!'}); 
+			function confirmAction(){
+			      var confirmed = confirm("Are you sure? This will remove this entry.");
+			      return confirmed;
+			}
 		</script>
 	</body>
 </html>
